+++ project/src/components/container/container.tsx
@@ -1,20 +1,29 @@
-import { useState, useEffect } from 'react';
+import { useEffect, useState } from 'react';
 import { useLocation } from 'react-router';
 import { useAppDispatch } from '../../hooks';
 import { resetFilmsCount } from '../../store/actions';
 import { removeLastSlash } from '../../utils';
+import { SYMBOLS } from '../../const';
 
-function ResetFilmsCount(): JSX.Element | null {
+function Container({ children }: { children: React.ReactNode }): JSX.Element {
+  // scroll to top
+  const location = useLocation();
+
+  useEffect(() => {
+    window.scrollTo(0, 0);
+  }, [location]);
+
+  // reset films count
   const dispatch = useAppDispatch();
   const [currentPath, setCurrentPath] = useState<string>();
   const [resetStatus, setResetStatus] = useState<boolean>(false);
   const pathname = removeLastSlash(useLocation().pathname);
 
-  if (currentPath === undefined) {
+  if (!currentPath && currentPath !== SYMBOLS.EMPTY) {
     setCurrentPath(removeLastSlash(pathname));
   }
 
-  if (currentPath !== undefined && currentPath !== pathname) {
+  if (currentPath !== pathname) {
     setCurrentPath(pathname);
     setResetStatus(true);
   }
@@ -24,7 +33,8 @@ function ResetFilmsCount(): JSX.Element | null {
     setResetStatus(false);
   }, [dispatch, resetStatus]);
 
-  return null;
+  return <div>{children}</div>;
 }
 
-export default ResetFilmsCount;
+export default Container;
+
