+++ project/src/components/films-list/films-list.tsx
@@ -4,8 +4,9 @@ import { MORE_LIKE_THIS_MAX_COUNT } from '../../const';
 import { useAppSelector } from '../../hooks';
 
 function FilmsList({films, isMoreLikeThis, isMyList}: FilmsListProps): JSX.Element {
-  const maxToShow = useAppSelector((state) => isMyList ? films.length : state.maxToShow);
-  const filmsToShow = isMoreLikeThis ? films.slice(0, MORE_LIKE_THIS_MAX_COUNT) : films.slice(0, maxToShow);
+  const stateMaxToShow = useAppSelector((state) => state.maxToShow);
+  const maxToShow = isMyList ? films.length : stateMaxToShow;
+  const filmsToShow = films.slice(0, isMoreLikeThis ? MORE_LIKE_THIS_MAX_COUNT : maxToShow);
 
   return (
     <div className="catalog__films-list">
