--- project/src/components/scroll-to-top/scroll-to-top.tsx
@@ -1,15 +1,0 @@
-import { useEffect } from 'react';
-import { useLocation } from 'react-router';
-
-function ScrollToTop({ children }: { children: React.ReactNode }): JSX.Element {
-  const location = useLocation();
-
-  useEffect(() => {
-    window.scrollTo(0, 0);
-  }, [location]);
-
-  return <div>{children}</div>;
-}
-
-export default ScrollToTop;
-
