+++ project/src/pages/my-list-page/my-list-page.tsx
@@ -18,7 +18,7 @@ function MyListPage({films}: MyListPageProps): JSX.Element {
       <section className="catalog">
         <h2 className="catalog__title visually-hidden">Catalog</h2>
 
-        <FilmsList films={films} isMoreLikeThis={false} isMyList />
+        <FilmsList films={films} isMyList />
       </section>
 
       <Footer />
