+++ project/src/types/film-list-props-type.ts
@@ -2,8 +2,8 @@ import { Films } from './film-type';
 
 type FilmsListProps = {
   films: Films;
-  isMoreLikeThis: boolean;
-  isMyList: boolean;
+  isMoreLikeThis?: boolean;
+  isMyList?: boolean;
 }
 
 export type { FilmsListProps };
